Simplified and cleaned up honor calculation.
--HG--
branch : trunk

+++ src/server/game/BattleGrounds/BattleGroundMgr.cpp
@@ -1891,8 +1891,8 @@ void BattleGroundMgr::BuildBattleGroundListPacket(WorldPacket *data, const uint6
     uint32 win_arena = plr->GetRandomWinner() ? BG_REWARD_WINNER_ARENA_LAST : BG_REWARD_WINNER_ARENA_FIRST;
     uint32 loos_kills = plr->GetRandomWinner() ? BG_REWARD_LOOSER_HONOR_LAST : BG_REWARD_LOOSER_HONOR_FIRST;
 
-    win_kills = (uint32)Trinity::Honor::hk_honor_at_level(plr->getLevel(), win_kills);
-    loos_kills = (uint32)Trinity::Honor::hk_honor_at_level(plr->getLevel(), loos_kills);
+    win_kills = Trinity::Honor::hk_honor_at_level(plr->getLevel(), win_kills);
+    loos_kills = Trinity::Honor::hk_honor_at_level(plr->getLevel(), loos_kills);
 
     data->Initialize(SMSG_BATTLEFIELD_LIST);
     *data << uint64(guid);                                  // battlemaster guid
